Make TabNavV2 the new official version
