+++ common/views/components/TabNav/README.md
@@ -2,6 +2,4 @@
 
 To navigate through categorisations within a set of results.
 
-V2 currently exists to cover Concepts work, it will eventually overwrite the original version.
-
 [Edit this on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/common/views/components/TabNav/README.md)
